Create the custom hook useColaborators.js for the ColaboratorsApp.jsx
