+++ src/ColaboradoresApp.jsx
@@ -1,53 +1,21 @@
-import { useEffect, useState } from "react";
 import { Alert } from "./components/Alert";
 import { Buscador } from "./components/Buscador";
 import { Formulario } from "./components/Formulario";
 import { Listado } from "./components/Listado";
-import { getAllColaborators, searchColaborators } from "./services/colaboradores";
+import { useColaborators } from "./hooks/useColaborators";
 
 const ColaboradoresApp = () => {
-  const [colaboradores, setColaboradores] = useState([]);
 
-  const [searchResult, setSearchResult] = useState([]);
-
-  const [messages, setMessages] = useState([]);
-
-  const [searchValue, setSearchValue] = useState("");
-
-  useEffect(() => {
-    getData();
-  }, []);
-
-  const getData = () => {
-    const colaboradoresData = getAllColaborators();
-    setColaboradores([...colaboradoresData]);
-  };
-
-  const handleNewColaborador = (colaborador) => {
-    setColaboradores([...colaboradores, colaborador]);
-    setSearchResult([]);
-    setSearchValue("");
-  };
-
-  const handleValidationsErrors = (messages) => {
-    if (messages.length > 0) {
-      setMessages(messages);
-    } else {
-      setMessages([{
-        message: "Colaborador agregado con éxito!",
-        color: "success",
-      }]);
-    }
-  };
-
-  const searchColaborador = (searchValue) => {
-    const result = searchColaborators(colaboradores, searchValue)
-    setSearchResult(result);
-  };
-
-  const dataToTable = () => {
-    return searchValue ? searchResult : colaboradores;
-  };
+  const {
+    searchValue,
+    searchResult,
+    messages,
+    searchColaborador,
+    setSearchValue,
+    dataToTable, 
+    handleNewColaborador,
+    handleValidationsErrors
+  } = useColaborators();
 
   return (
     <div className="container">
@@ -79,13 +47,13 @@ const ColaboradoresApp = () => {
           />
           {
             messages.length > 0
-              && messages.map((error) => (
-                <Alert
-                  key={self.crypto.randomUUID()}
-                  message={error.message}
-                  color={error.color}
-                />
-              ))
+            && messages.map((error) => (
+              <Alert
+                key={self.crypto.randomUUID()}
+                message={error.message}
+                color={error.color}
+              />
+            ))
           }
         </div>
       </div>
